+++ src/rend/color/mod.rs
@@ -3,6 +3,9 @@
 //! Chromatic functionality.
 //
 
+// safety:
+#![cfg_attr(feature = "safe_color", forbid(unsafe_code))]
+
 mod base;
 mod error;
 mod namespace;