+++ src/rend/font/mod.rs
@@ -3,6 +3,9 @@
 //! Font functionality.
 //
 
+// safety:
+#![cfg_attr(feature = "safe_font", forbid(unsafe_code))]
+
 mod error;
 pub use error::*;