+++ src/rend/layout/mod.rs
@@ -3,6 +3,9 @@
 //! Layout functionality.
 //
 
+// safety:
+#![cfg_attr(feature = "safe_layout", forbid(unsafe_code))]
+
 mod error;
 pub use error::*;