+++ src/sys/io/define_no_std_io/fns.rs
@@ -1,9 +1,6 @@
 // devela::sys::io::reimplement_no_std::fns
 
-use crate::{
-    mem::MaybeUninit,
-    sys::{IoErrorKind, IoRead, IoResult, IoWrite},
-};
+use crate::{sf, IoErrorKind, IoRead, IoResult, IoWrite};
 
 /// Copies the entire contents of a reader into a writer.
 ///
@@ -14,23 +11,19 @@ where
     R: ?Sized + IoRead,
     W: ?Sized + IoWrite,
 {
-    #[cfg(not(feature = "unsafe_array"))]
+    #[cfg(any(feature = "safe_io", not(feature = "unsafe_array")))]
     let mut buf = [0u8; S];
-    #[cfg(feature = "unsafe_array")]
-    let mut buf = MaybeUninit::<[u8; S]>::uninit();
+    #[cfg(all(not(feature = "safe_io"), feature = "unsafe_array"))]
+    let mut buf = crate::MaybeUninit::<[u8; S]>::uninit();
     let mut written = 0;
 
     loop {
         let len = match reader.read({
-            #[cfg(not(feature = "unsafe_array"))]
-            {
-                &mut buf
-            }
-            #[cfg(feature = "unsafe_array")]
+            #[cfg(any(feature = "safe_io", not(feature = "unsafe_array")))]
+            sf! { &mut buf }
+            #[cfg(all(not(feature = "safe_io"), feature = "unsafe_array"))]
             // SAFETY: The read method will initialize the portion of the buffer it writes to.
-            unsafe {
-                &mut *buf.as_mut_ptr()
-            }
+            sf! { unsafe { &mut *buf.as_mut_ptr() } }
         }) {
             Ok(0) => return Ok(written),
             Ok(len) => len,
@@ -38,9 +31,9 @@ where
             Err(e) => return Err(e),
         };
 
-        #[cfg(not(feature = "unsafe_array"))]
+        #[cfg(any(feature = "safe_io", not(feature = "unsafe_array")))]
         writer.write_all(&buf[..len])?;
-        #[cfg(feature = "unsafe_array")]
+        #[cfg(all(not(feature = "safe_io"), feature = "unsafe_array"))]
         // SAFETY: `buf` has been initialized up to `len` bytes by the `read` operation.
         writer.write_all(unsafe { &buf.assume_init_ref()[..len] })?;