+++ src/sys/io/mod.rs
@@ -8,20 +8,19 @@
 #![doc = include_str!("./define_no_std_io/MODIFICATIONS.md")]
 //
 
-// warnings:
-#![allow(unused_imports)]
+// safety:
+#![cfg_attr(feature = "safe_io", forbid(unsafe_code))]
 
 #[cfg(not(feature = "std"))]
-#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "sys")))]
-mod define_no_std_io;
+crate::items! {
+    mod define_no_std_io;
+    pub use define_no_std_io::*;
+}
 #[cfg(feature = "std")]
-#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "sys")))]
-mod reexport_std;
-
-#[cfg(not(feature = "std"))]
-pub use define_no_std_io::*;
-#[cfg(feature = "std")]
-pub use reexport_std::*;
+crate::items! {
+    mod reexport_std;
+    pub use reexport_std::*;
+}
 
 pub(crate) mod all {
     #[doc(inline)]